# contracts/openapi/backend-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garanti Kulucka Backend API",
    "version": "0.1.0",
    "description": "Backend-owned HTTP contract consumed by the web app."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/settings/{key}": {
      "put": {
        "operationId": "upsertSetting",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setting"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/integrations/accounts": {
      "get": {
        "operationId": "listIntegrationAccounts",
        "responses": {
          "200": {
            "description": "Integration accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationAccountList"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "upsertIntegrationAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertIntegrationAccountRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored integration account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationAccount"
                }
              }
            }
          }
        }
      }
    },
    "/admin/integrations/accounts/{account_public_id}": {
      "get": {
        "operationId": "getIntegrationAccountSnapshot",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountPublicId"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration account snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationAccountSnapshot"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/integrations/accounts/{account_public_id}/analytics-summary": {
      "get": {
        "operationId": "getInstagramAnalyticsSummary",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountPublicId"
          }
        ],
        "responses": {
          "200": {
            "description": "Instagram analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstagramAnalyticsSummary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/admin/integrations/accounts/{account_public_id}/settings/{key}": {
      "put": {
        "operationId": "upsertIntegrationSetting",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountPublicId"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertIntegrationSettingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored integration setting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSetting"
                }
              }
            }
          }
        }
      }
    },
    "/admin/integrations/accounts/{account_public_id}/tokens/{token_type}": {
      "put": {
        "operationId": "upsertIntegrationToken",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountPublicId"
          },
          {
            "name": "token_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertIntegrationTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stored integration token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationToken"
                }
              }
            }
          }
        }
      }
    },
    "/admin/integrations/provider-attempts": {
      "get": {
        "operationId": "listProviderAttempts",
        "parameters": [
          {
            "name": "provider_key",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "account_public_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider attempt audit rows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAttemptList"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/integrations/provider-cron-triggers/{provider_key}": {
      "post": {
        "operationId": "triggerProviderCronDebug",
        "parameters": [
          {
            "name": "provider_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ptt",
                "surat"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCronTriggerRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Fixture-safe provider cron trigger attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAttempt"
                }
              }
            }
          }
        }
      }
    },
    "/admin/integrations/instagram-publish-previews": {
      "post": {
        "operationId": "createInstagramPublishPreview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstagramPublishPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Fixture-safe Instagram publish preview attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAttempt"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
...
      "AccountPublicId": {
        "name": "account_public_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "DataEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array"
          }
        }
      },
...
      "IntegrationAccountList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DataEnvelope"
          }
        ]
      },
      "ProviderAttemptList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DataEnvelope"
          }
        ]
      },
...
      "ProviderCronTriggerRequest": {
        "type": "object",
        "required": [
          "idempotency_key"
        ],
        "properties": {
          "idempotency_key": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "InstagramPublishPreviewRequest": {
        "type": "object",
        "required": [
          "image_url",
          "caption",
          "idempotency_key"
        ],
        "properties": {
          "account_public_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_url": {
            "type": "string",
            "format": "uri"
          },
          "caption": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2200
          },
          "idempotency_key": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "ProviderAttempt": {
        "type": "object",
        "required": [
          "public_id",
          "provider_key",
          "request_id",
          "operation",
          "direction",
          "status",
          "duration_ms",
          "retry_decision",
          "started_at",
          "updated_at"
        ],
        "properties": {
          "public_id": {
            "type": "string"
          },
          "provider_key": {
            "type": "string"
          },
          "account_public_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_id": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "status_code": {
            "type": [
              "integer",
              "null"
            ]
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0
          },
          "retry_decision": {
            "type": "string"
          },
          "next_retry_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "idempotency_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_metadata": {},
          "provider_request_preview": {},
          "response_metadata": {},
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpsertSettingRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": true,
          "scope": {
            "type": "string",
            "default": "global"
          },
          "is_secret": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "Setting": {
        "type": "object",
        "required": [
          "key",
          "scope",
          "value",
          "is_secret",
          "updated_at"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "value": true,
          "is_secret": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "UpsertIntegrationAccountRequest": {
        "type": "object",
        "required": [
          "provider_key",
          "display_name"
        ],
        "properties": {
          "provider_key": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "external_account_id": {
            "type": [
              "string",
              "null"
            ],
            "default": null
          },
          "metadata": {
            "type": "object",
            "default": {}
          }
        }
      },
      "IntegrationAccount": {
        "type": "object",
        "required": [
          "public_id",
          "provider_key",
          "display_name"
        ],
        "properties": {
          "public_id": {
            "type": "string"
          },
          "provider_key": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "external_account_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "IntegrationAccountSnapshot": {
        "type": "object",
        "required": [
          "account",
          "settings",
          "tokens"
        ],
        "properties": {
          "account": {
            "$ref": "#/components/schemas/IntegrationAccount"
          },
          "settings": {
            "type": "array"
          },
          "tokens": {
            "type": "array"
          }
        }
      },
      "InstagramAnalyticsSummary": {
        "type": "object",
        "required": [
          "followers",
          "reach",
          "impressions",
          "profile_views",
          "engagement_rate"
        ],
        "properties": {
          "followers": {
            "type": "integer",
            "minimum": 0
          },
          "reach": {
            "type": "integer",
            "minimum": 0
          },
          "impressions": {
            "type": "integer",
            "minimum": 0
          },
          "profile_views": {
            "type": "integer",
            "minimum": 0
          },
          "engagement_rate": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "UpsertIntegrationSettingRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": true,
          "is_secret": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "IntegrationSetting": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Setting"
          }
        ]
      },
      "UpsertIntegrationTokenRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": true,
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "default": null
          }
        }
      },
      "IntegrationToken": {
        "type": "object",
        "required": [
          "token_type",
          "value",
          "expires_at"
        ],
        "properties": {
          "token_type": {
            "type": "string"
          },
          "value": true,
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}